Day/Night Cycle
The only important thing is the DayNightCycle script. Attach it to the
light source and the light source will rotate, simulating sunlight.
Unity comes with a built in sun that rotates with the light. Hopefully
we can change it to a minecraft like square at a later date. Also comes
with a is night boolean that can be used to activate things that happen
only at night or during the day, such as mob spawns.