Working Terrain Gen
Working Terrain Gen! To "finish" the game and put everything together, I
think (as far as I know...) all that needs to be done is to tie the
inventory system to the spawning system used here (and to add more
blocks to this system by making more block scripts.) This should be
achievable fairly easily by telling the player block placement in the
"Modify" script to place different blocks depending on what is selected
in the inventory. Same thing with removing blocks, the Modify script
removes blocks by replacing them with the air block and this can be set
up to only work on certain blocks depending on what tool is selected.
Also, it would probably be a good idea to merge the "modify" script with
the player controller.

Also again, the "Notes" script serves as a readme along with providing
several vague explanations on complicated things...