Day Night Spawn Cycle
A unity package containing a spawner that spawns random prefabs, which
ones depending on whether it is day or night. The spawners are meant to
be placed slightly above the ground, which should work because the
general shape of the ground is the same each time.

Also, if my day night cycle script from the first day was merged into
something else the new script will have to be modified slightly to point
at it.