4:pm work on spawning and AI
AI is 80% where it should be. (so more or less good enough to put in the
game) however I gtg now and don't have any more time to work on this for
now. If someone can get the zombies to climb over obstacles like the
pigs do it would be awesome.